pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sam-agent"
version = "0.0.1"
description = "Stateful Agent Model - Production-oriented AI agent system"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "SAM Team"}
]

dependencies = [
  "langgraph>=0.0.50",
  "langchain>=0.1.0",
  "pydantic>=2.0",
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
  "qdrant-client>=1.14,<2.0",
  "openai-whisper>=20230314",
  "TTS>=0.22.0",
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0"
]
[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.0.290",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
]

[tool.setuptools.packages.find]
include = ["sam_agent"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=sam_agent --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
